[camera] Combine `CameraModule` with `CameraViewManager`
