+++ packages/expo-camera/android/src/main/java/expo/modules/camera/CameraViewModule.kt
@@ -3,8 +3,11 @@ package expo.modules.camera
 import android.Manifest
 import android.content.Context
 import com.google.android.cameraview.AspectRatio
+import com.google.android.cameraview.Size
 import expo.modules.camera.tasks.ResolveTakenPictureAsyncTask
+import expo.modules.core.interfaces.services.UIManager
 import expo.modules.core.utilities.EmulatorUtilities
+import expo.modules.interfaces.barcodescanner.BarCodeScannerSettings
 import expo.modules.interfaces.permissions.Permissions
 import expo.modules.kotlin.Promise
 import expo.modules.kotlin.exception.Exceptions
@@ -12,10 +15,10 @@ import expo.modules.kotlin.functions.Queues
 import expo.modules.kotlin.modules.Module
 import expo.modules.kotlin.modules.ModuleDefinition
 
-class CameraModule : Module() {
+class CameraViewModule : Module() {
   override fun definition() = ModuleDefinition {
-    // TODO(@lukmccall): combine with the `CameraViewManager` and rename to `ExponentCamera`
-    Name("ExponentCameraModule")
+    Name("ExponentCamera")
+
     Constants(
       "Type" to mapOf(
         "front" to com.google.android.cameraview.Constants.FACING_FRONT,
@@ -170,6 +173,81 @@ class CameraModule : Module() {
         promise,
         Manifest.permission.RECORD_AUDIO
       )
+    }
+
+    View(ExpoCameraView::class) {
+      Events(
+        "onCameraReady",
+        "onMountError",
+        "onBarCodeScanned",
+        "onFacesDetected",
+        "onFaceDetectionError",
+        "onPictureSaved"
+      )
+
+      OnViewDestroys<ExpoCameraView> { view ->
+        val uiManager = appContext.legacyModule<UIManager>()
+        uiManager?.unregisterLifecycleEventListener(view)
+        view.cameraView.stop()
+      }
+
+      Prop("type") { view: ExpoCameraView, type: Int ->
+        view.cameraView.facing = type
+      }
+
+      Prop("ratio") { view: ExpoCameraView, ratio: String? ->
+        if (ratio == null) {
+          return@Prop
+        }
+        view.cameraView.setAspectRatio(AspectRatio.parse(ratio))
+      }
+
+      Prop("flashMode") { view: ExpoCameraView, torchMode: Int ->
+        view.cameraView.flash = torchMode
+      }
+
+      Prop("autoFocus") { view: ExpoCameraView, autoFocus: Boolean ->
+        view.cameraView.autoFocus = autoFocus
+      }
+
+      Prop("focusDepth") { view: ExpoCameraView, depth: Float ->
+        view.cameraView.focusDepth = depth
+      }
+
+      Prop("zoom") { view: ExpoCameraView, zoom: Float ->
+        view.cameraView.zoom = zoom
+      }
+
+      Prop("whiteBalance") { view: ExpoCameraView, whiteBalance: Int ->
+        view.cameraView.whiteBalance = whiteBalance
+      }
+
+      Prop("pictureSize") { view: ExpoCameraView, size: String? ->
+        if (size == null) {
+          return@Prop
+        }
+        view.cameraView.pictureSize = Size.parse(size)
+      }
+
+      Prop("barCodeScannerSettings") { view: ExpoCameraView, settings: Map<String, Any?>? ->
+        view.setBarCodeScannerSettings(BarCodeScannerSettings(settings))
+      }
+
+      Prop("useCamera2Api") { view: ExpoCameraView, useCamera2Api: Boolean ->
+        view.cameraView.setUsingCamera2Api(useCamera2Api)
+      }
+
+      Prop("barCodeScannerEnabled") { view: ExpoCameraView, barCodeScannerEnabled: Boolean ->
+        view.setShouldScanBarCodes(barCodeScannerEnabled)
+      }
+
+      Prop("faceDetectorEnabled") { view: ExpoCameraView, faceDetectorEnabled: Boolean ->
+        view.setShouldDetectFaces(faceDetectorEnabled)
+      }
+
+      Prop("faceDetectorSettings") { view: ExpoCameraView, settings: Map<String, Any>? ->
+        view.setFaceDetectorSettings(settings)
+      }
     }
   }
