+++ packages/expo-camera/expo-module.config.json
@@ -5,6 +5,6 @@
     "modules": ["CameraViewModule"]
   },
   "android": {
-    "modules": ["expo.modules.camera.CameraModule", "expo.modules.camera.CameraViewModule"]
+    "modules": ["expo.modules.camera.CameraViewModule"]
   }
 }
