+++ packages/expo-camera/src/ExponentCameraManager.ts
@@ -1,6 +1,5 @@
 import { NativeModulesProxy } from 'expo-modules-core';
 
-const CameraManager: Record<string, any> =
-  NativeModulesProxy.ExponentCamera || NativeModulesProxy.ExponentCameraModule;
+const CameraManager: Record<string, any> = NativeModulesProxy.ExponentCamera;
 
 export default CameraManager;
